requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
ax-platform==0.3.4
beartype==0.17.2
botorch==0.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
deepdiff==6.7.1
executing==2.0.1
filelock==3.14.0
fonttools==4.52.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
GenProcTrees==0.0.1.dev1
gpytorch==1.11
greenlet==3.0.3
grpcio==1.64.0
h5py==3.10.0
hyperopt==0.2.7
idna==3.7
image-similarity-measures==0.3.6
imageio==2.34.1
ipython==8.24.0
ipywidgets==8.1.3
jaxtyping==0.2.28
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.11
k-Wave-python==0.3.3
kiwisolver==1.4.5
lazy_loader==0.4
linear-operator==0.5.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.0.8
multipledispatch==1.0.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
phasepack==1.5
pillow==10.3.0
plotly==5.22.0
prompt_toolkit==3.0.45
protobuf==5.27.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==16.1.0
pyFFTW==0.13.1
Pygments==2.18.0
pyparsing==3.1.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
pyre-extensions==0.0.30
pyro-api==0.1.2
pyro-ppl==1.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
ray==2.23.0
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
scikit-image==0.20.0
scikit-learn==1.5.0
scipy==1.12.0
six==1.16.0
SQLAlchemy==1.4.17
stack-data==0.6.3
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.5.22
tk==0.1.0
torch==2.3.0
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.11
